# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Producto MusicPro",
        "description": "Documentacion de como usar la API Bodega",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/producto": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Listado de todos los productos",
                "operationId": "0ae2747cdd1920ea98ae53f68e021678",
                "responses": {
                    "200": {
                        "description": "Peticion hecha correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "Nombre": {
                                                        "type": "string",
                                                        "example": "Bateria"
                                                    },
                                                    "Codigo": {
                                                        "type": "string",
                                                        "example": "sdh2DF23FF"
                                                    },
                                                    "Descripcion": {
                                                        "type": "string",
                                                        "example": "Color: Blanco"
                                                    },
                                                    "Precio": {
                                                        "type": "number",
                                                        "example": "2500"
                                                    },
                                                    "Stock": {
                                                        "type": "number",
                                                        "example": "4"
                                                    },
                                                    "Foto": {
                                                        "type": "string",
                                                        "example": "https://picsum.photos/200/300"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Producto"
                ],
                "summary": "Registrar la información de un producto",
                "operationId": "31a377e56d699faab358339be17a750e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombres": {
                                                "type": "string"
                                            },
                                            "apellidos": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombres": "Aderson Felix",
                                    "apellidos": "Jara Lazaro"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombres": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "apellidos": {
                                            "type": "string",
                                            "example": "Jara Lazaro"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/producto/{id}": {
            "get": {
                "tags": [
                    "Producto"
                ],
                "summary": "Mostrar la información de un producto",
                "operationId": "af41299eb66ce2d7e6a7c81112895d23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Bateria"
                                        },
                                        "Codigo": {
                                            "type": "string",
                                            "example": "KSDH22D2"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Producto] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Producto"
                ],
                "summary": "Actualizar la información de un Producto",
                "operationId": "e295410d9e6c7e46dd5cb53161d676ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "nombres": {
                                                "type": "string"
                                            },
                                            "apellidos": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombres": "Aderson Felix Editado",
                                    "apellidos": "Jara Lazaro Editado"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombres": {
                                            "type": "string",
                                            "example": "Aderson Felix Editado"
                                        },
                                        "apellidos": {
                                            "type": "string",
                                            "example": "Jara Lazaro Editado"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The apellidos field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Producto"
                ],
                "summary": "Eliminar la información de un Producto",
                "operationId": "26f9b1a91c32db076a712a31ed7ddd2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}